Separate WinRT and COM call planners
Keep WinRT signatures and metadata isolated while Classic COM owns its method registry and lowers through a private native-call backend. Add exact struct validation, preserve ABI-compatible array projections, and document the runtime boundaries.

Copilot-Session: 41a826d4-b04a-4932-b1b8-31173fc5ab40
