+++ crates/dynwinrt/src/lib.rs
@@ -7,6 +7,7 @@ mod abi;
 mod call;
 pub mod com;
 mod interfaces;
+mod native_call;
 mod result;
 mod roapi;
 mod signature;
